Refactored code and double checked the code
