+++ backend/src/app.js
@@ -2,11 +2,7 @@ import express from "express";
 const app = express();
 app.use(express.json());
 
-import cors from "cors";
-app.use(cors());
-
 import mongoose from "mongoose";
-
 mongoose
   .connect("mongodb://localhost:27017/swapi-app")
   .then(() => console.log("Connected to MongoDB..."))
