--- backend/src/data.json
@@ -1,7 +1,0 @@
-[
-    { "_id": 1, "name": "Ymannen" },
-    { "_id": 2, "name": "Xmannen" },
-    { "_id": 3, "name": "Emperor Benno" },
-    { "_id": 4, "name": "Darth Skontan" },
-    { "_id": 5, "name": "Stormtrooper Noel" }
-]
