+++ frontend/src/components/index.ts
@@ -1,0 +1,7 @@
+export { default as Background } from "./Background";
+export { default as LogoContainer } from "./LogoContainer";
+export { default as Logo } from "./Logo";
+export { default as CollectionButton } from "./CollectionButton";
+export { default as MusicPlayer } from "./MusicPlayer";
+export { default as CharacterForm } from "./CharacterForm";
+export { default as CharacterList } from "./CharacterList";
