+++ frontend/src/hooks/useCharacters.ts
@@ -6,7 +6,6 @@ const useCharacters = () => {
   const [characters, setCharacters] = useState<Character[]>([]);
   const [error, setError] = useState("");
   const [isLoading, setIsLoading] = useState(false);
-  const [hasLoaded, setHasLoaded] = useState(false);
 
   useEffect(() => {
     setIsLoading(true);
@@ -15,19 +14,17 @@ const useCharacters = () => {
       .then((res) => {
         setCharacters(res.data);
         setIsLoading(false);
-        setHasLoaded(true);
       })
       .catch((error) => {
         if (error instanceof CanceledError) return;
         setError(error.message);
         setIsLoading(false);
-        setHasLoaded(false);
       });
 
     return () => cancel();
-  }, [hasLoaded]);
+  }, []);
 
-  return { characters, setCharacters, error, setError, isLoading };
+  return { characters, setCharacters, error, isLoading };
 };
 
 export default useCharacters;
